+++ src/components/form/form.tsx
@@ -72,6 +72,7 @@ export class Form {
 
   createFormPage(label?: string, formFacet?: string) {
     let page = document.createElement('gl-form-page');
+    page.formFacet = formFacet;
     page.facets = this.filter(this._schema.facets, formFacet);
     page.fields = this.filter(this._schema.fields, formFacet);
     page.root = !formFacet;